Attendance model

+++ WebAPI/WebAPI/DTOs/EmployeeDTO.cs
@@ -7,7 +7,7 @@ namespace WebAPI.DTOs
     public class EmployeeDTO
     {
         public int Id { get; set; }
-        public int SSN { get; set; }
+        public double SSN { get; set; }
         public string FullName { get; set; }
         public string Address { get; set; }
         public string PhoneNumber { get; set; }
