+++ WebAPI/WebAPI/Models/Employee.cs
@@ -8,7 +8,7 @@ namespace WebAPI.Models
         [Key]
         public int Id { get; set; }
 
-        public int SSN { get; set; }
+        public double SSN { get; set; }
         public string FullName { get; set; }
         public string Address { get; set; }
         public string PhoneNumber { get; set; }
